% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ext.med_Circ.R
\name{Ext.med_Circ}
\alias{Ext.med_Circ}
\title{Ext.med_Circ}
\usage{
Ext.med_Circ(X)
}
\arguments{
\item{X}{sample data on unit circle}
}
\value{
geometric median on unit circle
}
\description{
Ext.med_Circ
}
