% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ext.med_Shape.R
\name{Ext.med_Shape}
\alias{Ext.med_Shape}
\title{Ext.med_Shape}
\usage{
Ext.med_Shape(X)
}
\arguments{
\item{X}{sample data on shape space}
}
\value{
geometric median of planar shape
}
\description{
Ext.med_Shape
}
