% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi_RELR.R
\name{Multi_RELR}
\alias{Multi_RELR}
\title{Multi_RELR}
\usage{
Multi_RELR(grid, X, Response, H)
}
\arguments{
\item{X}{Covariate}

\item{Response}{shape (Matrix N*K)}

\item{H}{bandwidths (column)}

\item{gird}{Evaluation point}
}
\value{
Est estimated values

h optimal bandwidth

RMSE with observed
}
\description{
Multi_RELR
}
