% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uni_RELR.R
\name{Uni_RELR}
\alias{Uni_RELR}
\title{Uni_RELR}
\usage{
Uni_RELR(grid, X, Response, H)
}
\arguments{
\item{X}{Covariate}

\item{Response}{shape (Matrix N*K)}

\item{H}{bandwidths (column)}

\item{gird}{Evaluation point}
}
\value{
h optimal bandwidth

RMSE with observed

RMSE_True with true value
}
\description{
Uni_RELR
}
